riscv

+++ gccbuild/x86_64-w64-mingw32/riscv64-linux-gnu.sh
@@ -676,11 +676,15 @@ if [[ $isnativebuild != "yes" ]]; then
 	fi
 	if [[ ${FREESTANDINGBUILD} != "yes" ]]; then
 		TOOLCHAINS_BUILD=$TOOLCHAINS_BUILD TOOLCHAINSPATH=$TOOLCHAINSPATH GMPMPFRMPCHOST=$HOST GMPMPFRMPCBUILD=${currentpath}/targetbuild/$HOST GMPMPFRMPCPREFIX=$SYSROOT/usr $relpath/buildgmpmpfrmpc.sh
-
+		if [ $? -ne 0 ]; then
+			echo "$HOST gmp mpfr mpc build failed"
+			exit 1
+		fi	
 		if [ ! -f ${currentpath}/targetbuild/$HOST/gcc_phase2/.copysysrootprefixsuccess ]; then
 			rm -rf $PREFIX/sysroot
 			mkdir -p $PREFIX/sysroot
 			cp -r --preserve=links $SYSROOT/* $PREFIX/sysroot/
+			exit 1
 			echo "$(date --iso-8601=seconds)" > ${currentpath}/targetbuild/$HOST/gcc_phase2/.copysysrootprefixsuccess
 		fi
 	fi
